+++ oceanics-io-api/src/auth.spec.ts
@@ -1,19 +1,16 @@
 import fetch from "node-fetch";
 import { describe, expect, test } from '@jest/globals';
-import { API_PATH, fetchToken, Authorization, register, apiFetch } from "./shared/middleware.spec";
+import { API_PATH, fetchToken, Authorization, register, apiFetch } from "../test-utils";
 
 const AUTH_PATH = `${API_PATH}/auth`;
 
 /**
- * Stand alone tests for the Auth flow. Includes initial
+ * Stand alone tests for Auth flow. Includes initial
  * teardown of test artifacts remaining in the graph.
- *
- * On a clean database, the first test will fail.
  */
 describe("auth", function () {
-
   /**
-   * Check the required environment variables.
+   * Check required environment variables.
    */
   describe("environment", function () {
     test.concurrent.each([
@@ -28,19 +25,19 @@ describe("auth", function () {
     });
   })
 
-  let expectRegistrationToFail = true;
-
   /**
    * Isolate destructive actions so that it can be called
    * with mocha grep flag.
    */
   describe("teardown", function () {
-    test("clears non-provider, nodes", async function () {
+    /**
+     * Fails on empty database.
+     */
+    test("clears non-provider nodes", async function () {
       const token = await fetchToken();
       const response = await apiFetch(token, AUTH_PATH, "DELETE")();
       expect(response.status).toBe(204);
-      expectRegistrationToFail = false;
-    }, 5000);
+    });
   });
 
   /**
@@ -49,7 +46,7 @@ describe("auth", function () {
    */
   describe("register", function () {
 
-    (expectRegistrationToFail ? test.skip : test.concurrent)("allows registration with valid API key", async function () {
+    test("allows registration with valid API key", async function () {
       const response = await register(process.env.SERVICE_PROVIDER_API_KEY??"");
       expect(response.status).toEqual(200);
     });
@@ -70,23 +67,23 @@ describe("auth", function () {
    */
   describe("login", function () {
 
-    test.concurrent("prevents duplicate registration", async function () {
+    test("prevents duplicate registration", async function () {
       const response = await register(process.env.SERVICE_PROVIDER_API_KEY??"");
       expect(response.status).toEqual(403);
     });
 
-    test.concurrent("returns well-formed token", async function () {
+    test("returns well-formed token", async function () {
       const token = await fetchToken();
       expect(typeof token).toBe("string");
       expect(token).not.toBeFalsy();
     });
 
-    test.concurrent("denies missing header with 403", async function () {
+    test("denies missing header with 403", async function () {
       const response = await fetch(AUTH_PATH);
       expect(response.status).toEqual(403);
     });
 
-    test.concurrent("denies wrong credentials with 403", async function () {
+    test("denies wrong credentials with 403", async function () {
       const response = await fetch(AUTH_PATH, {
         headers: {
           Authorization: Authorization(undefined, "a-very-bad-password", undefined),
@@ -95,7 +92,7 @@ describe("auth", function () {
       expect(response.status).toEqual(403);
     });
 
-    test.concurrent("denies wrong salt with 403", async function () {
+    test("denies wrong salt with 403", async function () {
       const response = await fetch(AUTH_PATH, {
         headers: {
           Authorization: Authorization(undefined, undefined, "a-very-bad-secret"),
